ml_model/requirements.txt
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
flask>=3.0.0
flask-cors>=4.0.0
joblib>=1.3.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langgraph>=0.1.0
python-dotenv>=1.0.0
